pipelines/base-dit-pipeline-experiments/torch_requirements.txt
--index-url https://download.pytorch.org/whl/cu124
torch==2.4.1+cu124 --find-links https://download.pytorch.org/whl/cu124
torchvision==0.19.1+cu124 
